backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.12
python-dotenv==1.0.1
langchain==0.3.7
langchain-groq==0.2.1
langchain-community==0.3.7
langchain-text-splitters==0.3.2
chromadb==0.5.18
sentence-transformers==3.3.0
pypdf==5.1.0
python-docx==1.1.2
pydantic==2.9.2
pydantic-settings==2.6.1
onnxruntime>=1.14.1
